requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.5
django-celery==3.2.1
django-cities==0.5.0.5
django-cors-headers==2.1.0
django-filter==1.1.0
django-mathfilters==0.4.0
django-model-utils==3.0.0
django-rest-multiple-models==1.8.2
django-widget-tweaks==1.4.1
djangorestframework==3.7.1
et-xmlfile==1.0.1
gunicorn==19.7.1
idna==2.6
jdcal==1.3
kombu==3.0.37
lxml==4.0.0
openpyxl==2.4.9
psycopg2==2.7.3.1
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
six==1.11.0
static3==0.7.0
swapper==1.1.0
tqdm==4.17.1
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
